% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_import_module.R
\name{classify_geo_assay}
\alias{classify_geo_assay}
\title{Classify a GEO object's assay type}
\usage{
classify_geo_assay(obj, gse_accession = "")
}
\arguments{
\item{obj}{A GEO ExpressionSet or GSE object}

\item{gse_accession}{Character. Used only for error messages.}
}
\description{
Returns one of: "bulk_rnaseq", "microarray", "scrna", "excluded", "unknown"
}
