% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_import_module.R
\name{classify_supp_file}
\alias{classify_supp_file}
\title{Classify a supplementary file by assay type}
\usage{
classify_supp_file(fname)
}
\description{
Returns "rnaseq", "salmon", "scrna", "excluded", or "unknown"
}
