% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score_VariableAssociation.R
\name{create_contrast_column}
\alias{create_contrast_column}
\title{Create Contrast Column in Metadata}
\usage{
create_contrast_column(metadata, variable_name, contrast)
}
\arguments{
\item{metadata}{A data frame containing sample metadata.}

\item{variable_name}{A character string specifying the column name in
\code{metadata} that represents the variable of interest.}

\item{contrast}{A character string representing the contrast in the form
"(A + B) - (C + D)" (e.g.).}
}
\value{
A subset of \code{metadata} with an added \code{cohentest} column, indicating
group membership based on the contrast.
}
\description{
This function extracts and processes contrast groups from a given contrast
string, then assigns contrast labels to a metadata subset based on the
variable of interest.
}
\keyword{internal}
