% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{deps}
\alias{deps}
\alias{use_marker}
\alias{useMarker}
\title{Dependencies}
\usage{
use_marker()

useMarker()
}
\description{
Include marker dependencies at the top of your shiny UI.
}
\examples{
library(shiny)

ui <- fluidPage(
 useMarker()
)

}
