% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_import_module.R
\name{detect_superseries}
\alias{detect_superseries}
\title{Detect if a GEO accession is a SuperSeries}
\usage{
detect_superseries(gse_accession)
}
\description{
Returns a character vector of sub-series GSE IDs, or character(0) if not.
}
