% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_import_module.R
\name{.extract_expr_matrix}
\alias{.extract_expr_matrix}
\title{Internal: extract numeric expression matrix from a data.frame}
\usage{
.extract_expr_matrix(df, fname = "")
}
\description{
Internal: extract numeric expression matrix from a data.frame
}
