% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_import_module.R
\name{.pick_best_excel_sheet}
\alias{.pick_best_excel_sheet}
\title{Internal: pick best Excel sheet by name scoring}
\usage{
.pick_best_excel_sheet(sheets)
}
\description{
Internal: pick best Excel sheet by name scoring
}
