% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_import_module.R
\name{extract_geo_metadata}
\alias{extract_geo_metadata}
\title{Extract tidy metadata from a GEO ExpressionSet}
\usage{
extract_geo_metadata(obj)
}
\arguments{
\item{obj}{GEO ExpressionSet}
}
\value{
data.frame: rows = samples, columns = attributes (SampleID always first)
}
\description{
Extract tidy metadata from a GEO ExpressionSet
}
