% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_import_module.R
\name{geoImportServer}
\alias{geoImportServer}
\title{GEO Import Module — Server}
\usage{
geoImportServer(id, log_fn = NULL)
}
\arguments{
\item{id}{Shiny module namespace ID.}

\item{session}{Parent Shiny session (for notifications).}
}
\value{
A list of reactives: \code{expr_data}, \code{meta_data}, \code{status}.
}
\description{
GEO Import Module — Server
}
