% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lorem}
\alias{lorem}
\title{Lorem Ipsum}
\format{An object of class \code{character} of length 1.}
\source{
stringi package
}
\usage{
lorem
}
\description{
A lorem ipsum paragraph
}
\keyword{datasets}
