% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateScores_ssGSEA.R
\name{CalculateScores_ssGSEA}
\alias{CalculateScores_ssGSEA}
\title{Calculate Gene Signature Scores using ssGSEA}
\usage{
CalculateScores_ssGSEA(data, metadata = NULL, gene_sets)
}
\arguments{
\item{data}{A data frame of normalized (non-transformed) counts where each row is a gene and each column is a sample.}

\item{metadata}{A data frame containing sample metadata (optional).}

\item{gene_sets}{A named list of gene signatures.}
}
\value{
A list of data frames containing ssGSEA scores for each signature.
}
\description{
Computes an enrichment score for each gene signature in each sample using the single-sample Gene Set Enrichment Analysis (ssGSEA).
}
\keyword{internal}
