% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counts_example}
\alias{counts_example}
\title{Gene Expression Counts for Marthandan et al. (2016) RNA-Seq Data}
\format{
A numeric matrix with rows corresponding to genes and columns corresponding to samples.
Row names indicate gene symbols and column names are sample identifiers.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE63577}
}
\usage{
counts_example
}
\description{
A numeric matrix containing gene expression data from the Marthandan et al. (2016) study (GEO code GSE63577).
The data represent normalized (non-transformed) counts for genes across samples.
}
\details{
The samples were quality-controlled using FastQC and aligned using kallisto (v0.46.1) against the
RefSeq v109 transcriptome. The original dataset had 45 samples, but intermediate time points from
HFF and MRC5 were removed.
}
\references{
Marthandan S, Priebe S, Baumgart M, Groth M et al. Similarities in Gene Expression Profiles during
In Vitro Aging of Primary Human Embryonic Lung and Foreskin Fibroblasts. Biomed Res Int 2015;2015:731938. PMID: 26339636

Marthandan S, Baumgart M, Priebe S, Groth M et al. Conserved Senescence Associated Genes and Pathways
in Primary Human Fibroblasts Detected by RNA-Seq. PLoS One 2016;11(5):e0154531. PMID: 27140416
}
\keyword{datasets}
