% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_CohenD.R
\name{flatten_results}
\alias{flatten_results}
\title{Flatten a Nested List of Results into a Data Frame}
\usage{
flatten_results(nested_list)
}
\arguments{
\item{nested_list}{A nested list with structure: \code{list(method = list(signature = data.frame(...)))}.}
}
\value{
A data frame combining all the nested data frames, with added columns \code{method} and \code{signature}.
}
\description{
Converts a nested list (where the first level is a method, the second level is a gene signature,
and the third level is a data frame) into a single data frame. Additional columns for method and signature
are added to the data frame.
}
\keyword{internal}
