% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lorem}
\alias{lorem}
\title{Lorem Ipsum}
\format{
A vector of length one comprised of 616 characters forming lorem ipsum.
}
\source{
stringi package
}
\usage{
data(lorem)
}
\description{
A lorem ipsum paragraph
}
\keyword{datasets}
