% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score_VariableAssociation.R
\name{remove_division}
\alias{remove_division}
\title{Remove Division Notation in Contrast Labels}
\usage{
remove_division(contrasts)
}
\arguments{
\item{contrasts}{A character vector containing contrast labels.}
}
\value{
A character vector with division notation removed.
}
\description{
This function removes division notation (e.g., \verb{/2}, \verb{/3}) after closing parentheses in contrast labels.
}
