% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_similarity.R
\name{signature_similarity}
\alias{signature_similarity}
\title{Plot Signature Similarity via Jaccard Index}
\usage{
signature_similarity(
  signatures,
  other_user_signatures = NULL,
  collection = NULL,
  subcollection = NULL,
  limits = c(0, 1),
  title_size = 12,
  color_values = c("#F9F4AE", "#B44141"),
  title = NULL,
  num_sigs_toplot = NULL,
  jaccard_threshold = 0,
  msig_subset = NULL
)
}
\arguments{
\item{signatures}{A named list of character vectors representing reference gene signatures.}

\item{other_user_signatures}{Optional. A named list of character vectors representing other user-defined signatures to compare against. If \code{NULL}, MSigDB gene sets (if provided) will be used.}

\item{collection}{Optional. MSigDB collection name (e.g., \code{"H"} for hallmark, \code{"C2"} for curated gene sets). Use msigdbr::msigdbr_collections() for the available options.}

\item{subcollection}{Optional. Subcategory within an MSigDB collection (e.g., \code{"CP:REACTOME"}). Use msigdbr::msigdbr_collections() for the available options.}

\item{limits}{Numeric vector of length 2. Limits for the Jaccard index color scale. Default is \code{c(0, 1)}.}

\item{title_size}{Integer specifying the font size for the plot title. Default is \code{12}.}

\item{color_values}{Character vector of colors used in the Jaccard fill gradient. Default is \code{c("#F9F4AE", "#B44141")}.}

\item{title}{Optional. Custom title for the plot. If \code{NULL}, the title defaults to \code{"Signature Overlap"}.}

\item{num_sigs_toplot}{Optional. Integer. Maximum number of comparison signatures (including user and MSigDB) to display.}

\item{jaccard_threshold}{Numeric. Minimum Jaccard index required for an MSigDB gene set to be included in the plot. Default is \code{0}.}

\item{msig_subset}{Optional. Character vector of MSigDB gene set names to subset from the specified collection. Useful to restrict analysis to a specific set of pathways. If supplied, other filters will apply only to this subset.}
}
\value{
A \code{ggplot} object showing a heatmap of Jaccard similarities between reference and comparison signatures.
}
\description{
Visualises the similarity between user-defined gene signatures and either
other user-defined signatures or MSigDB gene sets, using the Jaccard index.
Produces a heatmap of pairwise Jaccard indices between reference and comparison
signatures.
}
\examples{
sig1 <- list(A = c("TP53", "BRCA1", "EGFR"))
sig2 <- list(B = c("TP53", "MYC", "EGFR"), C = c("GATA3", "STAT3"))

signature_similarity(
  signatures = sig1,
  other_user_signatures = sig2
)

}
