% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualiseIndividualGenes.R
\name{VisualiseIndividualGenes}
\alias{VisualiseIndividualGenes}
\title{Visualize Individual Genes' Expression Using Various Methods}
\usage{
VisualiseIndividualGenes(
  data,
  metadata,
  genes,
  method = c("violins", "expr_heatmap", "corr_heatmap", "roc", "effectsize", "all"),
  plot = TRUE,
  GroupingVariable = NULL,
  ColorVariable = NULL,
  ColorValues = NULL,
  pointSize = 2,
  title = NULL,
  widthTitle = 16,
  y_limits = NULL,
  legend_nrow = NULL,
  ncol = NULL,
  nrow = NULL,
  divide = NULL,
  invert_divide = FALSE,
  xlab = NULL,
  colorlab = NULL
)
}
\arguments{
\item{data}{A data frame where rows correspond to genes and columns correspond to samples.
The row names should be gene identifiers.}

\item{metadata}{A data frame containing sample metadata. The first column should contain
sample identifiers matching the column names of \code{data}.}

\item{genes}{A character vector of gene names to be plotted. If more than 20 genes are provided, a warning is issued.}

\item{method}{A character string specifying which visualization method to use.
Options are \code{"violins"}, \code{"expr_heatmap"}, \code{"corr_heatmap"}, \code{"roc"}, \code{"effectsize"}, or \code{"all"}.
Default is \code{"violins"}. If \code{"all"} is selected, all methods will be applied.}

\item{plot}{A logical value indicating whether to display the generated plots (\code{TRUE}) or not (\code{FALSE}).
Default is \code{TRUE}.}

\item{GroupingVariable}{A character string specifying the column name in \code{metadata} to be used
for grouping samples on the x-axis in violin plots.}

\item{ColorVariable}{A character string specifying the column name in \code{metadata} used
to color the points in violin plots.}

\item{ColorValues}{An optional named vector mapping unique values of \code{ColorVariable} to specific colors.}

\item{pointSize}{A numeric value specifying the size of jittered points in violin plots. Default is \code{2}.}

\item{title}{An optional character string specifying the title of the plot.}

\item{widthTitle}{An optional integer specifying the maximum width of the title before inserting line breaks. Default is \code{16}.}

\item{y_limits}{A numeric vector of length 2 specifying the limits of the y-axis in violin plots.
If \code{NULL}, the y-axis is adjusted automatically. Default is \code{NULL}.}

\item{legend_nrow}{A numeric value specifying the number of rows in the legend for violin plots.
If \code{NULL}, the default number of rows is determined automatically.}

\item{ncol}{An optional numeric value specifying the number of columns in the grid layout for violin plots.
If \code{NULL}, it is automatically determined.}

\item{nrow}{An optional numeric value specifying the number of rows in the grid layout for violin plots.
If \code{NULL}, it is automatically determined.}

\item{divide}{A character string specifying a column in \code{metadata} used to divide the violin plots into facets.
Default is \code{NULL}.}

\item{invert_divide}{A logical value indicating whether to invert the facet layout of violin plots.
Default is \code{FALSE}.}

\item{xlab}{An optional character string specifying the label for the x-axis in violin plots.
Default is the name of \code{GroupingVariable}.}

\item{colorlab}{An optional character string specifying the legend title for the color variable.
Default is \code{NULL}.}
}
\value{
A named list where each selected method (e.g., \code{"violins"}, \code{"expr_heatmap"}, etc.) is a key.
Each element in this list is itself a named list with the following structure:
\describe{
\item{\code{data}}{A data frame or matrix containing the transformed data used for plotting, formatted according to the selected method.}
\item{\code{plot}}{The generated plot object (e.g., a ggplot2 object) corresponding to the visualization.}
\item{\code{aux}}{Additional parameters, specific of each method.}
}
If \code{plot = TRUE}, the plots are also displayed in the output.
}
\description{
This function provides multiple visualization methods for analyzing individual genes
from an expression dataset. Users can generate violin plots, expression heatmaps,
correlation heatmaps, ROC curves, and effect size plots.
}
\details{
This function calls multiple sub-functions based on the selected visualization method:
\itemize{
\item \code{IndividualGenes_Violins()}: Generates violin plots for gene expression (log2 transformed).
\item \code{IndividualGenes_ExprHeatmap()}: Creates a heatmap of expression values.
\item \code{IndividualGenes_CorrHeatmap()}: Computes and visualizes a correlation heatmap.
\item \code{IndividualGenes_ROC()}: Plots receiver operating characteristic (ROC) curves.
\item \code{IndividualGenes_EffectSize()}: Computes and visualizes effect sizes.
}

The function returns a list of plots. If \code{method = "all"}, all available visualizations are included.
}
